+++ src/primitive/defs.rs
@@ -690,14 +690,13 @@ primitive!(
     ///   : > □"banana" □"orange"
     ///   : > □"banana"  "orange"
     ///
-    /// Most non-pervasive monadic functions, like [reverse], [first], [transpose], etc, will work on box elements without needing to [un][box] them.
-    /// Unlike pervasive functions, [box] depth is only preserved for some of these.
-    /// ex: ⊢□□[1 2 3]
+    /// For non-pervasive functions, boxed arrays need to be [un][box]ed before they can be operated on.
+    /// ex! ⊢  □[1 2 3]
+    /// ex: ⊢ °□[1 2 3]
+    /// [under][un][box] is useful when you want to re-[box] the result.
     /// ex: $ Reverse these words
     ///   : ⊜□≠@ .
-    ///   : ∵⇌.
-    ///
-    /// For more complex operations, you can use [under][un][box].
+    ///   : ∵⍜°□⇌.
     /// ex: {"Hey" "there" "world"}
     ///   : ≡⍜°□(⊂⊢.)
     /// ex: PrepLen ← $"_ _"⧻.